Add skeleton for library project
This commit introduces a new command `library`. Invoking `ember library`, a directory structure for an Ember.js library is created. It differs to the app layout that it just has the essential files. Also, the Rakefile has tasks to release a new version as well as uploading a version to GitHub downloads